추가 : admin & pagination

+++ src/users/admin/admin.controller.ts
@@ -4,26 +4,59 @@ import {
   UseGuards,
   Delete,
   Param,
+  Req,
+  Query,
   ForbiddenException,
+  UnauthorizedException,
 } from '@nestjs/common';
 import { Roles } from './roles.decorator';
 import { RolesGuard } from './roles.guard';
 import { UsersService } from '../users.service';
 
-@Controller('admin')
-@UseGuards(RolesGuard)
-export class UsersController {
+@Controller('admin') // @UseGuards(RolesGuard)
+export class AdminController {
   constructor(private readonly usersService: UsersService) {}
 
   @Get('user-list')
-  @Roles(['admin'])
-  async getUserList() {
-    return this.usersService.getUserList();
+  async getUserList(
+    @Req() request,
+    @Query('page') page = 1,
+    @Query('perPage') perPage = 10,) {
+    const accessToken = request.headers.authorization;
+
+    if (!accessToken) {
+      throw new UnauthorizedException('토큰이 없습니다');
+    }
+
+    const userId = await this.usersService.verifyAccessTokenAndGetUserId(
+      accessToken,
+    );
+
+    const user = await this.usersService.getUser(userId);
+
+    if (user.roles !== 'admin') {
+      throw new ForbiddenException('admin 계정이 아닙니다.');
+    }
+
+    return this.usersService.getUserList(page, perPage);
   }
 
-  @Delete('users/:user_id')
-  @Roles(['admin'])
-  async deleteUser(@Param('user_id') user_id: string) {
+  @Delete(':user_id')
+  async deleteUser(@Param('user_id') user_id, @Req() request) {
+    const accessToken = request.headers.authorization;
+
+    if (!accessToken) {
+      throw new UnauthorizedException('토큰이 없습니다');
+    }
+
+    const userId = await this.usersService.verifyAccessTokenAndGetUserId(
+      accessToken,
+    );
+    const user = await this.usersService.getUser(userId);
+
+    if (user.roles !== 'admin') {
+      throw new ForbiddenException('admin 계정이 아닙니다.');
+    }
     return this.usersService.deleteUser(user_id);
   }
 }
